First iteration of the manual instrumentation
