+++ Sources/DatadogSDKTesting/DDTestObserver.swift
@@ -10,36 +10,15 @@ import Foundation
 @_implementationOnly import SigmaSwiftStatistics
 @_implementationOnly import XCTest
 
-internal class DDTestObserver: NSObject, XCTestObservation {
-    var tracer: DDTracer
-
+class DDTestObserver: NSObject, XCTestObservation {
     static let testNameRegex = try! NSRegularExpression(pattern: "([\\w]+) ([\\w]+)", options: .caseInsensitive)
     static let supportsSkipping = NSClassFromString("XCTSkippedTestContext") != nil
     static let tracerVersion = (Bundle(for: DDTestObserver.self).infoDictionary?["CFBundleShortVersionString"] as? String) ?? "unknown"
-    var currentBundleName = ""
-    var currentBundleFunctionInfo = FunctionMap()
-    var currentTestExecutionOrder = 0
-    var initialProcessId = Int(ProcessInfo.processInfo.processIdentifier)
-    var codeOwners: CodeOwners?
 
-    var rLock = NSRecursiveLock()
-    private var privateCurrentTestSpan: Span?
-    var currentTestSpan: Span? {
-        get {
-            rLock.lock()
-            defer { rLock.unlock() }
-            return privateCurrentTestSpan
-        }
-        set {
-            rLock.lock()
-            defer { rLock.unlock() }
-            privateCurrentTestSpan = newValue
-        }
-    }
-
-    init(tracer: DDTracer) {
+    weak var ddTest: DDTest!
+    init(ddTest: DDTest) {
+        self.ddTest = ddTest
         XCUIApplication.swizzleMethods
-        self.tracer = tracer
         super.init()
     }
 
@@ -48,24 +27,13 @@ internal class DDTestObserver: NSObject, XCTestObservation {
     }
 
     func testBundleWillStart(_ testBundle: Bundle) {
-        currentBundleName = testBundle.bundleURL.deletingPathExtension().lastPathComponent
-
-        #if !os(tvOS) && (targetEnvironment(simulator) || os(macOS))
-        DDSymbolicator.createDSYMFileIfNeeded(forImageName: currentBundleName)
-        currentBundleFunctionInfo = FileLocator.testFunctionsInModule(currentBundleName)
-        #endif
-        if let workspacePath = tracer.env.workspacePath {
-            codeOwners = CodeOwners(workspacePath: URL(fileURLWithPath: workspacePath))
-        }
-
-        if !tracer.env.disableCrashHandler {
-            DDCrashes.install()
-        }
+        let bundleName = testBundle.bundleURL.deletingPathExtension().lastPathComponent
+        ddTest.bundleStart(name: bundleName)
     }
 
     func testBundleDidFinish(_ testBundle: Bundle) {
         /// We need to wait for all the traces to be written to the backend before exiting
-        tracer.flush()
+        ddTest.bundleEnd()
     }
 
     func testCaseWillStart(_ testCase: XCTestCase) {
@@ -78,109 +46,32 @@ internal class DDTestObserver: NSObject, XCTestObservation {
         let testSuite = String(testCase.name[suiteRange])
         let testName = String(testCase.name[nameRange])
 
-        currentTestExecutionOrder = currentTestExecutionOrder + 1
-
-        let attributes: [String: String] = [
-            DDGenericTags.type: DDTagValues.typeTest,
-            DDGenericTags.resourceName: "\(currentBundleName).\(testSuite).\(testName)",
-            DDTestTags.testName: testName,
-            DDTestTags.testSuite: testSuite,
-            DDTestTags.testFramework: "XCTest",
-            DDTestTags.testBundle: currentBundleName,
-            DDTestTags.testType: DDTagValues.typeTest,
-            DDTestTags.testExecutionOrder: "\(currentTestExecutionOrder)",
-            DDTestTags.testExecutionProcessId: "\(initialProcessId)",
-            DDOSTags.osPlatform: tracer.env.osName,
-            DDOSTags.osArchitecture: tracer.env.osArchitecture,
-            DDOSTags.osVersion: tracer.env.osVersion,
-            DDDeviceTags.deviceName: tracer.env.deviceName,
-            DDDeviceTags.deviceModel: tracer.env.deviceModel,
-            DDRuntimeTags.runtimeName: "Xcode",
-            DDRuntimeTags.runtimeVersion: tracer.env.runtimeVersion,
-            DDTracerTags.tracerLanguage: "swift",
-            DDTracerTags.tracerVersion: DDTestObserver.tracerVersion
-        ]
-
-        let testSpan = tracer.startSpan(name: testCase.name, attributes: attributes)
-
-        // Is not a UITest until a XCUIApplication is launched
-        testSpan.setAttribute(key: DDTestTags.testIsUITest, value: false)
-
-        if !tracer.env.disableDDSDKIOSIntegration {
-            tracer.addPropagationsHeadersToEnvironment()
-        }
-
-        let functionName = testSuite + "." + testName
-        if let functionInfo = currentBundleFunctionInfo[functionName] {
-            var filePath = functionInfo.file
-            if let workspacePath = tracer.env.workspacePath,
-               let workspaceRange = filePath.range(of: workspacePath + "/")
-            {
-                filePath.removeSubrange(workspaceRange)
-            }
-            testSpan.setAttribute(key: DDTestTags.testSourceFile, value: filePath)
-            testSpan.setAttribute(key: DDTestTags.testSourceStartLine, value: functionInfo.startLine)
-            testSpan.setAttribute(key: DDTestTags.testSourceEndLine, value: functionInfo.endLine)
-            if let owners = codeOwners?.ownersForPath(filePath) {
-                testSpan.setAttribute(key: DDTestTags.testCodeowners, value: owners)
-            }
-        }
-
-        tracer.env.addTagsToSpan(span: testSpan)
-
-        if let testSpan = testSpan as? RecordEventsReadableSpan {
-            let simpleSpan = SimpleSpanData(spanData: testSpan.toSpanData())
-            DDCrashes.setCustomData(customData: SimpleSpanSerializer.serializeSpan(simpleSpan: simpleSpan))
-        }
-        currentTestSpan = testSpan
+        ddTest.start(name: testName, testSuite: testSuite)
     }
 
     func testCaseDidFinish(_ testCase: XCTestCase) {
-        guard let activeTest = currentTestSpan else {
-            return
-        }
-        var status: String
-        if DDTestObserver.supportsSkipping, testCase.testRun?.hasBeenSkipped == true {
-            status = DDTagValues.statusSkip
-            activeTest.status = .ok
-        } else if testCase.testRun?.hasSucceeded ?? false {
-            status = DDTagValues.statusPass
-            activeTest.status = .ok
-        } else {
-            status = DDTagValues.statusFail
-            activeTest.status = .error(description: "Test failed")
-        }
+        addBenchmarkTagsIfNeeded(testCase: testCase)
 
-        activeTest.setAttribute(key: DDTestTags.testStatus, value: status)
-        addBenchmarkTagsIfNeeded(testCase: testCase, activeTest: activeTest)
-        activeTest.end()
-        tracer.backgroundWorkQueue.sync {}
-        currentTestSpan = nil
-        DDTestMonitor.instance?.networkInstrumentation?.endAndCleanAliveSpans()
+        if DDTestObserver.supportsSkipping, testCase.testRun?.hasBeenSkipped == true {
+            ddTest.end(status: .skip)
+        } else if testCase.testRun?.hasSucceeded ?? false {
+            ddTest.end(status: .pass)
+        } else {
+            ddTest.end(status: .fail)
+        }
     }
 
     #if swift(>=5.3)
     func testCase(_ testCase: XCTestCase, didRecord issue: XCTIssue) {
-        guard let activeTest = currentTestSpan else {
-            return
-        }
-        activeTest.setAttribute(key: DDTags.errorType, value: AttributeValue.string(issue.compactDescription))
-        activeTest.setAttribute(key: DDTags.errorMessage, value: AttributeValue.string(issue.description))
-        if let detailedDescription = issue.detailedDescription {
-            activeTest.setAttribute(key: DDTags.errorStack, value: AttributeValue.string(detailedDescription))
-        }
+        ddTest.setErrorInfo(type: issue.compactDescription, message: issue.description, callStack: issue.detailedDescription)
     }
     #else
     func testCase(_ testCase: XCTestCase, didFailWithDescription description: String, inFile filePath: String?, atLine lineNumber: Int) {
-        guard let activeTest = currentTestSpan else {
-            return
-        }
-        activeTest.setAttribute(key: DDTags.errorType, value: AttributeValue.string(description))
-        activeTest.setAttribute(key: DDTags.errorMessage, value: AttributeValue.string("test_failure: \(filePath ?? ""):\(lineNumber)"))
+        ddTest.testSetErrorInfo(type: description, message: "test_failure: \(filePath ?? ""):\(lineNumber)", callStack: nil)
     }
     #endif
 
-    private func addBenchmarkTagsIfNeeded(testCase: XCTestCase, activeTest: Span) {
+    private func addBenchmarkTagsIfNeeded(testCase: XCTestCase) {
         guard let metricsForId = testCase.value(forKey: "_perfMetricsForID") as? [XCTPerformanceMetric: AnyObject],
               let metric = metricsForId.first(where: {
                   let measurements = $0.value.value(forKey: "measurements") as? [Double]
@@ -194,45 +85,7 @@ internal class DDTestObserver: NSObject, XCTestObservation {
             return
         }
 
-        activeTest.setAttribute(key: DDTestTags.testType, value: DDTagValues.typeBenchmark)
         let values = measurements.map { $0 * 1_000_000_000 } // Convert to nanoseconds
-        activeTest.setAttribute(key: DDBenchmarkTags.benchmarkRuns, value: values.count)
-        activeTest.setAttribute(key: DDBenchmarkTags.statisticsN, value: values.count)
-        if let average = Sigma.average(values) {
-            activeTest.setAttribute(key: DDBenchmarkTags.durationMean, value: average)
-        }
-        if let max = Sigma.max(values) {
-            activeTest.setAttribute(key: DDBenchmarkTags.statisticsMax, value: max)
-        }
-        if let min = Sigma.min(values) {
-            activeTest.setAttribute(key: DDBenchmarkTags.statisticsMin, value: min)
-        }
-        if let mean = Sigma.average(values) {
-            activeTest.setAttribute(key: DDBenchmarkTags.statisticsMean, value: mean)
-        }
-        if let median = Sigma.median(values) {
-            activeTest.setAttribute(key: DDBenchmarkTags.statisticsMedian, value: median)
-        }
-        if let stdDev = Sigma.standardDeviationSample(values) {
-            activeTest.setAttribute(key: DDBenchmarkTags.statisticsStdDev, value: stdDev)
-        }
-        if let stdErr = Sigma.standardErrorOfTheMean(values) {
-            activeTest.setAttribute(key: DDBenchmarkTags.statisticsStdErr, value: stdErr)
-        }
-        if let kurtosis = Sigma.kurtosisA(values) {
-            activeTest.setAttribute(key: DDBenchmarkTags.statisticsKurtosis, value: kurtosis)
-        }
-        if let skewness = Sigma.skewnessA(values) {
-            activeTest.setAttribute(key: DDBenchmarkTags.statisticsSkewness, value: skewness)
-        }
-        if let percentile99 = Sigma.percentile(values, percentile: 0.99) {
-            activeTest.setAttribute(key: DDBenchmarkTags.statisticsP99, value: percentile99)
-        }
-        if let percentile95 = Sigma.percentile(values, percentile: 0.95) {
-            activeTest.setAttribute(key: DDBenchmarkTags.statisticsP95, value: percentile95)
-        }
-        if let percentile90 = Sigma.percentile(values, percentile: 0.90) {
-            activeTest.setAttribute(key: DDBenchmarkTags.statisticsP90, value: percentile90)
-        }
+        ddTest.testSetBenchmarkInfo(measureName: "", measureUnit: "", values: values)
     }
 }
