+++ Sources/DatadogSDKTesting/DatadogSDKTesting.h
@@ -12,6 +12,4 @@ FOUNDATION_EXPORT double DatadogSDKTestingVersionNumber;
 //! Project version string for DatadogSDKTesting.
 FOUNDATION_EXPORT const unsigned char DatadogSDKTestingVersionString[];
 
-// In this header, you should import all the public headers of your framework using statements like #import <DatadogSDKTesting/PublicHeader.h>
-
 #import <DatadogSDKTesting/DDSymbolAddress.h>
