+++ Tests/DatadogSDKTesting/OutputCapture/StderrCaptureTests.swift
@@ -22,7 +22,7 @@ class StderrCaptureTests: XCTestCase {
         let capturer = StderrCapture()
 
         let span = tracer.startSpan(name: "Unnamed", attributes: [:]) as! RecordEventsReadableSpan
-        DDTestMonitor.instance?.testObserver?.currentTestSpan = span
+        DDTestMonitor.instance?.ddTest?.currentTestSpan = span
         capturer.stderrMessage(tracer: tracer, string: stringToCapture)
         Thread.sleep(forTimeInterval: 0.5)
         span.status = .ok
@@ -44,7 +44,7 @@ class StderrCaptureTests: XCTestCase {
 
         let date = Date()
         let span = tracer.startSpan(name: "Unnamed", attributes: [:], date: date) as! RecordEventsReadableSpan
-        DDTestMonitor.instance?.testObserver?.currentTestSpan = span
+        DDTestMonitor.instance?.ddTest?.currentTestSpan = span
         capturer.stderrMessage(tracer: tracer, string: stringToCapture)
         Thread.sleep(forTimeInterval: 0.6)
         span.status = .ok
